Add separate class: AncaTest.
